chore: rename modal folder, remove unused modal component
