+++ src/components/organisms/Modals/ShareYourCoughModal.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image'
 import Link from 'next/link'
-import VirufyLogo from '~/assets/static/images/logo/virufy.svg'
+import { useI18n } from '~/i18n'
 import Modal from './Modal'
 
 interface Prop {
@@ -9,14 +9,16 @@ interface Prop {
 }
 
 const ShareYourCoughModal = ({ isOpen, handleClose }: Prop) => {
+  const { ShareYourCough: ShareYourCough } = useI18n()
+
   return (
     <Modal isOpen={isOpen} handleClose={handleClose}>
       {/* Modal Container */}
       <div className="py-16 text-center font-medium">
-        {/* type cast to remove unsafe assignment error */}
+        {/* Image Container */}
         <div className="flex items-center justify-center">
           <Image
-            src={VirufyLogo as string}
+            src={ShareYourCough.image}
             alt="Virufy logo"
             width={140}
             height={50}
@@ -24,21 +26,19 @@ const ShareYourCoughModal = ({ isOpen, handleClose }: Prop) => {
         </div>
 
         {/* Text */}
-        <div className="my-10 text-sm text-white">
-          An Independent Nonprofit Research Organization
-        </div>
+        <div className="my-10 text-sm text-white">{ShareYourCough.text}</div>
         <h1 className="mb-16 text-2xl text-green-500">
-          Covid-19 Cough Data Collection Study
+          {ShareYourCough.title}
         </h1>
 
         {/* Link Container */}
         <div>
           <Link
-            href="https://virufy.org/study/welcome"
+            href={ShareYourCough.linkUrl}
             target="_blank"
             className="rounded-full bg-gradient-to-b from-green-500 to-blue-500 py-4 px-12 text-xl text-white"
           >
-            Lets Start!
+            {ShareYourCough.linkText}
           </Link>
         </div>
       </div>
